standalone ErrorList
