+++ siteupdate/cplusplus/siteupdate.cpp
@@ -49,6 +49,7 @@ class HGEdge;
 #include "classes/WaypointQuadtree/WaypointQuadtree.h"
 #include "classes/Route/Route.h"
 #include "classes/ConnectedRoute/ConnectedRoute.h"
+#include "classes/ErrorList/ErrorList.h"
 #include "classes/Waypoint/Waypoint.h"
 #include "classes/HighwaySegment/HighwaySegment.h"
 #include "classes/HighwaySystem/HighwaySystem.h"
@@ -67,7 +68,6 @@ class HGEdge;
 #include "functions/valid_num_str.cpp"
 #include "classes/DBFieldLength.cpp"
 #include "classes/ElapsedTime.cpp"
-#include "classes/ErrorList.cpp"
 #include "classes/Region.cpp"
 #include "classes/HighwaySystem/HighwaySystem.cpp"
 #include "classes/DatacheckEntry.cpp"
